+++ packages/react/src/components/List/ListItem.tsx
@@ -105,7 +105,6 @@ class ListItem extends UIComponent<ReactProps<ListItemProps>, ListItemState> {
     truncateContent: PropTypes.bool,
     truncateHeader: PropTypes.bool,
 
-    accessibility: PropTypes.func,
     onClick: PropTypes.func,
     onFocus: PropTypes.func,
   }